requirements.txt
# env
python-dotenv==1.0.0

# langchain
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-ollama==0.1.3

# embeddings & vector store
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# document loaders 
pypdf==4.0.0
docx2txt==0.8
